chore(shared): move thing to share
