+++ packages/twenty-shared/src/utils/__tests__/trim-and-remove-duplicated-whitespaces-from-object-string-properties.test.ts
@@ -1,6 +1,5 @@
-import { EachTestingContext } from 'twenty-shared/testing';
-
-import { trimAndRemoveDuplicatedWhitespacesFromObjectStringProperties } from 'src/utils/trim-and-remove-duplicated-whitespaces-from-object-string-properties';
+import { EachTestingContext } from '@/testing/types/EachTestingContext.type';
+import { trimAndRemoveDuplicatedWhitespacesFromObjectStringProperties } from '../trim-and-remove-duplicated-whitespaces-from-object-string-properties';
 
 type SanitizeObjectStringPropertiesTestCase = EachTestingContext<{
   input: Record<string, any>;
