+++ packages/twenty-shared/src/utils/trim-and-remove-duplicated-whitespaces-from-string.ts
